# Optimize Enum.join/2 with empty string